ft_fclamp uses floats instead of doubles

+++ libs/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: nneronin <nneronin@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 11:56:18 by nneronin          #+#    #+#             */
-/*   Updated: 2021/07/26 14:20:43 by nneronin         ###   ########.fr       */
+/*   Updated: 2021/09/17 17:46:07 by nneronin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef struct s_list
 	struct s_list	*prev;
 }					t_list;
 
-double				ft_fclamp(double x, double a, double b);
+float				ft_fclamp(float x, float a, float b);
 int					ft_clamp(int x, int a, int b);
 void				add_to_list(t_list **list, void *new_v, size_t size);
 double				ft_atof(const char *str);
